# capstone_project/docs/swagger.yaml
definitions:
...
  game.GameCategoryDTO:
    properties:
      id:
        example: 1
        type: integer
      imageUrl:
        type: string
      name:
        example: Strategy
        type: string
    type: object
  game.GameCreateDTO:
    properties:
      endTime:
        example: "2023-04-01T18:00:00Z"
        type: string
      gameCategoryId:
        example: 1
        type: integer
      maxMember:
        example: 10
        type: integer
      name:
        example: Chess Tournament
        type: string
      requiredSkills:
        items:
          type: integer
        type: array
      startTime:
        example: "2023-04-01T10:00:00Z"
        type: string
    required:
    - endTime
    - gameCategoryId
    - maxMember
    - name
    - startTime
    type: object
...
  game.GameRequestDTO:
    properties:
      id:
        example: 1
        type: integer
      status:
        example: pending
        type: string
      user:
        $ref: '#/definitions/game.UserDTO'
    type: object
...
  game.GameResponseDTO:
    properties:
      createdAt:
        example: "2023-03-30T12:00:00Z"
        type: string
      endTime:
        example: "2023-04-01T18:00:00Z"
        type: string
      gameCategory:
        $ref: '#/definitions/game.GameCategoryDTO'
      gameOwner:
        $ref: '#/definitions/game.UserDTO'
      gameRequests:
        items:
          $ref: '#/definitions/game.GameRequestDTO'
        type: array
      id:
        example: 1
        type: integer
      maxMember:
        example: 10
        type: integer
      name:
        example: Chess Tournament
        type: string
      requiredSkills:
        items:
          $ref: '#/definitions/game.RequiredSkillDTO'
        type: array
      startTime:
        example: "2023-04-01T10:00:00Z"
        type: string
      updatedAt:
        example: "2023-03-30T12:00:00Z"
        type: string
    type: object
  game.GameUpdateDTO:
    properties:
      categories:
        items:
          type: integer
        type: array
      endTime:
        example: "2023-04-01T19:00:00Z"
        type: string
      maxMember:
        example: 12
        type: integer
      name:
        example: Updated Chess Tournament
        type: string
      requiredSkills:
        items:
          type: integer
        type: array
      startTime:
        example: "2023-04-01T11:00:00Z"
        type: string
    type: object
  game.RequiredSkillCreateDTO:
    properties:
      skills:
        items:
          type: string
        type: array
    required:
    - skills
    type: object
  game.RequiredSkillDTO:
    properties:
      id:
        example: 1
        type: integer
      label:
        example: Problem Solving
        type: string
      name:
        example: Problem Solving
        type: string
      value:
        example: Problem Solving
        type: string
    type: object
  game.UserDTO:
    properties:
      avatarUrl:
        example: https://example.com/avatar.jpg
        type: string
      id:
        example: 1
        type: integer
      username:
        example: john_doe
        type: string
    type: object
  handlers.ErrorResponse:
    properties:
      message:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /games:
    get:
      consumes:
      - application/json
      description: Get a list of all games with pagination, including game category
        details, required skills, and game requests
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/game.GameResponseDTO'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: List all games with pagination
      tags:
      - games
    post:
      consumes:
      - application/json
      description: Create a new game with the provided details
      parameters:
      - default: Bearer <Add access token here>
        description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Game details
        in: body
        name: game
        required: true
        schema:
          $ref: '#/definitions/game.GameCreateDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/game.GameResponseDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new game
      tags:
      - games
  /games/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a game by its ID
      parameters:
      - default: Bearer <Add access token here>
        description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Game ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a game
      tags:
      - games
    get:
      consumes:
      - application/json
      description: Get details of a game by its ID, including game category, required
        skills, game owner, and game requests
      parameters:
      - description: Game ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/game.GameResponseDTO'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Get a game by ID
      tags:
      - games
    put:
      consumes:
      - application/json
      description: Update a game with the provided details
      parameters:
      - default: Bearer <Add access token here>
        description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Game ID
        in: path
        name: id
        required: true
        type: integer
      - description: Game details to update
        in: body
        name: game
        required: true
        schema:
          $ref: '#/definitions/game.GameUpdateDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/game.GameResponseDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a game
      tags:
      - games
...
  /required-skills:
    get:
      consumes:
      - application/json
      description: Get a list of all required skills from the database
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/game.RequiredSkillDTO'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: List all required skills
      tags:
      - required-skills
    post:
      consumes:
      - application/json
      description: Create multiple new required skills (requires user login)
      parameters:
      - default: Bearer <Add access token here>
        description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Required skills details
        in: body
        name: skills
        required: true
        schema:
          $ref: '#/definitions/game.RequiredSkillCreateDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            items:
              $ref: '#/definitions/game.RequiredSkillDTO'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create multiple new required skills
      tags:
      - required-skills
swagger: "2.0"